Fix scheduling by removing actively running process from queues.
Trust me, it's easier this way. If a process doesn't need to run
anymore, simply don't add it back to the scheduler. Otherwise,
maintaining the actively running pid between execs and halts is a chore
and causes more issues than it could possibly fix. If you need to get
the active process for any reason, get_proc_pcb is what you need to use.